Don't drop the user in #drop unless it exists

+++ lib/mysql_users/user.rb
@@ -25,6 +25,7 @@ module MysqlUsers
     end
 
     def drop
+      return unless exists?
       sql = "DROP USER #{user_address}"
       db_client.query(sql)
     end
